pyproject.toml
[tool.poetry]
name = "api_projector"
version = "0.1.0"
description = ""
authors = ["junte ltd"]

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "*"
psycopg2-binary = "*"
pymemcache = "*"
django = ">=4.1"
django-health-check = "*"
django-cors-headers = "*"
django_extensions = "*"
django-split-settings = "*"
django-model-utils = "*"
django-minio-storage = "*"
django-filter = "*"
djangorestframework = "*"
graphene-django = "*"
graphene-file-upload = "*"
drf-yasg = "*"
jnt-django-admin-tools = "*"
jnt-django-toolbox = "*"
jnt-django-graphene-toolbox = "*"
uWSGI = "*"
sentry-sdk = "*"
injector = "*"
ibmcloudant = "*"
redis = "*"
flower = "*"
fluent-logger = "*"
celery = "*"
pillow = "*"
types-PyYAML = "*"
social-auth-core = "*"
social-auth-app-django = "*"
PyYAML = "*"
pydantic = "*"

[tool.poetry.dev-dependencies]
mypy = "*"
polint = "*"
dennis = "*"
xenon = "*"
safety = "*"
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
pytest-xdist = "*"
httpretty = "*"
factory-boy = "*"
flake8 = "4.*" # wait for https://github.com/tholo/pytest-flake8/issues/87
flake8-pytest = "*"
flake8-django = "*"
flake8-pytest-style = "*"
flake8-class-attributes-order = "*"
wemake-python-styleguide = "*"
black = "*"
types-requests = "*"
types-python-dateutil = "*"

[tool.poetry.group.dev.dependencies]
py = "^1.11.0"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git          # root of the project
    | \.mypy_cache
    | \.pytest_cache
    | .*\/migrations
  )/
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
